\name{CNVRanges}
\docType{data}
\alias{CNVRanges}
\title{Genomic locations and indices of the simulated CNVs.}
\description{
      This data set gives the starts, ends, and the integer copy number of the
      simulated CNVs in the data set \code{\link{XRanges}} object.}
\usage{CNVRanges}
\format{A GRanges object with 20 rows and 40 value columns across 1 space.}
\source{http://www.bioinf.jku.at/cnmops/cnmops.html.}
\references{
Klambauer et al., cn.mops, \emph{Unpublished method.}, 2011
      }
\keyword{datasets}
