% Generated by roxygen2 (4.0.2): do not edit by hand
\name{normalizeChromosomes}
\alias{normalizeChromosomes}
\title{Normalization of NGS data.

Normalize quantitative NGS data in order to make counts comparable over
samples, i.e., correcting for different library sizes or coverages.
Scales each samples' reads such that the coverage is even for
all samples after normalization.}
\usage{
normalizeChromosomes(X, chr, normType = "median", qu = 0.25, ploidy)
}
\arguments{
\item{X}{Matrix of positive real values, where
columns are interpreted as samples and rows as genomic regions. An entry is
the read count of a sample in the genomic region. Alternatively this can be
a GRanges object containing the read counts as values.}

\item{chr}{Character vector that has as many elements as "X" has rows. The
vector assigns each genomic segment to a reference sequence (chromosome).}

\item{normType}{Type of the normalization technique. Each samples'
read counts
are scaled such that the total number of reads are comparable across
samples.
By this parameter one can decide to how the size factors are calculated
Possible choices are the minimal coverage
("min"), the mean or median coverage ("mean", "median") or any quantile
("quant"). If this parameter is set to the value "mode",
the read counts are scaled such that each samples'
most frequent value (the "mode") is equal after normalization.
Possible values are "mean","min","median","quant", and "mode".
Default = "median".}

\item{qu}{Real value between 0 and 1. Default = 0.25.}

\item{ploidy}{An integer value for each sample or each column in the read
count matrix. At least two samples must have a ploidy of 2. Default = "missing".}
}
\value{
A data matrix of normalized read counts with the same dimensions
as the input matrix X.
}
\description{
Normalization of NGS data.

Normalize quantitative NGS data in order to make counts comparable over
samples, i.e., correcting for different library sizes or coverages.
Scales each samples' reads such that the coverage is even for
all samples after normalization.
}
\examples{
data(cn.mops)
X.norm <- normalizeChromosomes(X)
}
\author{
Guenter Klambauer \email{klambauer@bioinf.jku.at}
}

